organise types into a dto folder (and sub folders)
