+++ server/src/dtos/discord/discord-channel.ts
