+++ server/src/dtos/discord/discord-log.ts
