+++ server/src/dtos/discord/discord-message-details.ts
