+++ server/src/dtos/discord/discord-message.ts
