+++ server/src/dtos/discord/formatted-discord-member.ts
