+++ server/src/dtos/discord/formatted-log-entry.ts
