+++ server/src/dtos/discord/index.ts
@@ -1,4 +1,3 @@
-export * from './auth-info';
 export * from './discord-action-type';
 export * from './discord-channel';
 export * from './discord-embed';
@@ -8,17 +7,6 @@ export * from './discord-message';
 export * from './discord-message-details';
 export * from './discord-message-post';
 export * from './discord-role';
-export * from './event-create-dto';
-export * from './event-dto';
-export * from './event-settings-dto';
-export * from './event-settings-upsert-dto';
 export * from './formatted-discord-member';
 export * from './formatted-log-entry';
-export * from './gw2-log-entry';
-export * from './gw2-member';
-export * from './gw2-rank';
 export * from './member-left-dto';
-export * from './recruitment-post-create-dto';
-export * from './recruitment-post-dto';
-export * from './warning-create-dto';
-export * from './warning-dto';
