+++ server/src/dtos/discord/member-left-dto.ts
