+++ server/src/dtos/events/event-dto.ts
