+++ server/src/dtos/events/event-settings-dto.ts
