+++ server/src/dtos/events/event-settings-upsert-dto.ts
