+++ server/src/dtos/events/index.ts
@@ -1,0 +1,4 @@
+export * from './event-create-dto';
+export * from './event-dto';
+export * from './event-settings-dto';
+export * from './event-settings-upsert-dto';
