+++ server/src/dtos/index.ts
@@ -1,0 +1,6 @@
+export * from './auth';
+export * from './discord';
+export * from './events';
+export * from './gw2';
+export * from './recruitment-post';
+export * from './warnings';
