+++ server/src/dtos/recruitment-post/index.ts
@@ -1,0 +1,2 @@
+export * from './recruitment-post-create-dto';
+export * from './recruitment-post-dto';
