+++ server/src/dtos/warnings/index.ts
@@ -1,0 +1,2 @@
+export * from './warning-create-dto';
+export * from './warning-dto';
